containerized docker front and back app

+++ book-network-front/src/app/services/token/token.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 import { JwtHelperService } from '@auth0/angular-jwt';
+import { isPlatformBrowser } from '@angular/common';
 
 
 @Injectable({
@@ -7,30 +8,41 @@ import { JwtHelperService } from '@auth0/angular-jwt';
 })
 export class TokenService {
 
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
 
   set token(token: string){
-    localStorage.setItem('token', token);
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem('token', token);
+    }
   }
 
-  get token(){
-    return localStorage.getItem('token') as string;
+  get token(): string | null {
+    if (isPlatformBrowser(this.platformId)) {
+      return localStorage.getItem('token');
+    }
+    return null;
   }
 
   isTokenNotValid(): boolean {
     return !this.isTokenValid();
   }
 
-  isTokenValid() : boolean{
-    const token:string = this.token;
-    if(!token){
-      return false
-    }
-    const jwtHelper: JwtHelperService = new JwtHelperService();
-    const isTokenExpired: boolean = jwtHelper.isTokenExpired(token);
-    if(isTokenExpired){
-      localStorage.clear();
+  isTokenValid(): boolean {
+    const token: string | null = this.token;
+    if (!token) {
       return false;
     }
+
+    const jwtHelper: JwtHelperService = new JwtHelperService();
+    const isTokenExpired: boolean = jwtHelper.isTokenExpired(token);
+
+    if (isTokenExpired) {
+      if (isPlatformBrowser(this.platformId)) {
+        localStorage.clear();
+      }
+      return false;
+    }
+
     return true;
   }
 }
